requirements.txt
dash
dash-bootstrap-components
folium
databricks-sql-connector
databricks-sdk
flask-login
pandas